Refactoring the component
