+++ ui/pages/confirm-transaction-base/confirm-transaction-base.component.js
@@ -1,8 +1,5 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-///: BEGIN:ONLY_INCLUDE_IN(flask)
-import { stripHexPrefix } from 'ethereumjs-util';
-///: END:ONLY_INCLUDE_IN
 import ConfirmPageContainer from '../../components/app/confirm-page-container';
 import TransactionDecoding from '../../components/app/transaction-decoding';
 import { isBalanceSufficient } from '../send/send.utils';
@@ -56,23 +53,14 @@ import {
 
 import { MIN_GAS_LIMIT_DEC } from '../send/send.constants';
 
-///: BEGIN:ONLY_INCLUDE_IN(flask)
-import { SnapInsight } from '../../components/app/confirm-page-container/flask/snap-insight';
-import { DropdownTab, Tab } from '../../components/ui/tabs';
-///: END:ONLY_INCLUDE_IN
-
-import {
-  NETWORK_TO_NAME_MAP,
-  ///: BEGIN:ONLY_INCLUDE_IN(flask)
-  CHAIN_ID_TO_NETWORK_ID_MAP,
-  ///: END:ONLY_INCLUDE_IN
-} from '../../../shared/constants/network';
+import { NETWORK_TO_NAME_MAP } from '../../../shared/constants/network';
 import {
   addHexes,
   hexToDecimal,
   hexWEIToDecGWEI,
 } from '../../../shared/modules/conversion.utils';
 import TransactionAlerts from '../../components/app/transaction-alerts';
+import TransactionInsights from '../../components/app/transaction-insights';
 
 const renderHeartBeatIfNotInTest = () =>
   process.env.IN_TEST ? null : <LoadingHeartBeat />;
@@ -158,9 +146,6 @@ export default class ConfirmTransactionBase extends Component {
     isMultiLayerFeeNetwork: PropTypes.bool,
     isBuyableChain: PropTypes.bool,
     isApprovalOrRejection: PropTypes.bool,
-    ///: BEGIN:ONLY_INCLUDE_IN(flask)
-    insightSnaps: PropTypes.arrayOf(PropTypes.object),
-    ///: END:ONLY_INCLUDE_IN
     assetStandard: PropTypes.string,
     useCurrencyRateCheck: PropTypes.bool,
   };
@@ -173,9 +158,6 @@ export default class ConfirmTransactionBase extends Component {
     editingGas: false,
     userAcknowledgedGasMissing: false,
     showWarningModal: false,
-    ///: BEGIN:ONLY_INCLUDE_IN(flask)
-    selectedInsightSnapId: this.props.insightSnaps[0]?.id,
-    ///: END:ONLY_INCLUDE_IN
   };
 
   componentDidUpdate(prevProps) {
@@ -318,12 +300,6 @@ export default class ConfirmTransactionBase extends Component {
   handleCloseEditGas() {
     this.setState({ editingGas: false });
   }
-
-  ///: BEGIN:ONLY_INCLUDE_IN(flask)
-  handleSnapSelected(snapId) {
-    this.setState({ selectedInsightSnapId: snapId });
-  }
-  ///: END:ONLY_INCLUDE_IN
 
   setUserAcknowledgedGasMissing() {
     this.setState({ userAcknowledgedGasMissing: true });
@@ -755,62 +731,7 @@ export default class ConfirmTransactionBase extends Component {
 
   ///: BEGIN:ONLY_INCLUDE_IN(flask)
   renderInsight() {
-    const { txData, insightSnaps } = this.props;
-    const { selectedInsightSnapId } = this.state;
-    const { txParams, chainId, origin } = txData;
-
-    const selectedSnap = insightSnaps.find(
-      ({ id }) => id === selectedInsightSnapId,
-    );
-
-    const allowedTransactionTypes =
-      txData.type === TransactionType.contractInteraction ||
-      txData.type === TransactionType.simpleSend ||
-      txData.type === TransactionType.tokenMethodSafeTransferFrom ||
-      txData.type === TransactionType.tokenMethodTransferFrom ||
-      txData.type === TransactionType.tokenMethodTransfer;
-
-    const networkId = CHAIN_ID_TO_NETWORK_ID_MAP[chainId];
-    const caip2ChainId = `eip155:${networkId ?? stripHexPrefix(chainId)}`;
-
-    if (!allowedTransactionTypes || !insightSnaps.length) {
-      return null;
-    }
-
-    const dropdownOptions = insightSnaps.map(
-      ({ id, manifest: { proposedName } }) => ({
-        value: id,
-        name: proposedName,
-      }),
-    );
-
-    return insightSnaps.length > 1 ? (
-      <DropdownTab
-        className="confirm-page-container-content__tab"
-        options={dropdownOptions}
-        selectedOption={selectedInsightSnapId}
-        onChange={(snapId) => this.handleSnapSelected(snapId)}
-      >
-        <SnapInsight
-          transaction={txParams}
-          origin={origin}
-          chainId={caip2ChainId}
-          selectedSnap={selectedSnap}
-        />
-      </DropdownTab>
-    ) : (
-      <Tab
-        className="confirm-page-container-content__tab"
-        name={selectedSnap.manifest.proposedName}
-      >
-        <SnapInsight
-          transaction={txParams}
-          origin={origin}
-          chainId={caip2ChainId}
-          selectedSnap={selectedSnap}
-        />
-      </Tab>
-    );
+    return <TransactionInsights txData={this.props.txData} />;
   }
   ///: END:ONLY_INCLUDE_IN
